move librarian agent
